+++ packages/commons/.env
@@ -24,7 +24,7 @@ ALG=RS256
 TYP=JWT
  
 M2M_TOKEN_ENDPOINT=https://auth.dev.interop.pagopa.it/token.oauth2
-M2M_PRIVATE_KEY_PATH=../client-anpr-residence-verification-mockup/keys/client-test-keypair.rsa.priv
+M2M_PRIVATE_KEY_PATH=mock_path/client-test-keypair.rsa.priv
 M2M_CLIENT_ID=6d61c86c-1519-4d0a-b0c3-94da5a33275a
 M2M_AUTH_AUDIENCE=auth.dev.interop.pagopa.it/client-assertion
 M2M_KID=FPg1nEE5d6N4J8dngfvBkywXXxS5DVPFyWXcGWnEj2g
